+++ source/op/pair_tab.cc
@@ -92,7 +92,8 @@ class PairTabOp : public OpKernel {
         context, (3 * nnei * nloc == rij_tensor.shape().dim_size(1)),
         errors::InvalidArgument("shape of rij should be 3 * nloc * nnei"));
     OP_REQUIRES(
-        context, (nnei * nloc == nlist_tensor.shape().dim_size(1)),
+        context,
+        (static_cast<int64_t>(nnei) * nloc == nlist_tensor.shape().dim_size(1)),
         errors::InvalidArgument("shape of nlist should be nloc * nnei"));
     OP_REQUIRES(context, (nloc == scale_tensor.shape().dim_size(1)),
                 errors::InvalidArgument("shape of scale should be nloc"));
