+++ source/op/prod_force_se_r_grad.cc
@@ -71,7 +71,9 @@ class ProdForceSeRGradOp : public OpKernel {
     OP_REQUIRES(
         context, (nloc * 3 == grad_shape.dim_size(1)),
         errors::InvalidArgument("input grad shape should be 3 x natoms"));
-    OP_REQUIRES(context, (nloc * ndescrpt * 3 == in_deriv_shape.dim_size(1)),
+    OP_REQUIRES(context,
+                (static_cast<int64_t>(nloc) * ndescrpt * 3 ==
+                 in_deriv_shape.dim_size(1)),
                 errors::InvalidArgument("number of descriptors should match"));
 
     // Create an output tensor
