+++ source/op/prod_virial.cc
@@ -79,7 +79,9 @@ class ProdVirialOp : public OpKernel {
     OP_REQUIRES(context,
                 (nloc * ndescrpt * 12 == in_deriv_tensor.shape().dim_size(1)),
                 errors::InvalidArgument("number of descriptors should match"));
-    OP_REQUIRES(context, (nloc * nnei * 3 == rij_tensor.shape().dim_size(1)),
+    OP_REQUIRES(context,
+                (static_cast<int64_t>(nloc) * nnei * 3 ==
+                 rij_tensor.shape().dim_size(1)),
                 errors::InvalidArgument("dim of rij should be nnei * 3"));
     OP_REQUIRES(context, (nnei == n_a_sel + n_r_sel),
                 errors::InvalidArgument("number of neighbors should match"));
