+++ source/op/soft_min_force_grad.cc
@@ -77,8 +77,10 @@ class SoftMinForceGradOp : public OpKernel {
     OP_REQUIRES(
         context, (nloc * 3 == grad_shape.dim_size(1)),
         errors::InvalidArgument("input grad shape should be 3 x natoms"));
-    OP_REQUIRES(context, (nloc * nnei * 3 == sw_deriv_shape.dim_size(1)),
-                errors::InvalidArgument("number of sw deriv should match"));
+    OP_REQUIRES(
+        context,
+        (static_cast<int64_t>(nloc) * nnei * 3 == sw_deriv_shape.dim_size(1)),
+        errors::InvalidArgument("number of sw deriv should match"));
     OP_REQUIRES(context, (nnei == n_a_sel + n_r_sel),
                 errors::InvalidArgument("number of neighbors should match"));
